fix: rebase resolve conflicts
