+++ src/containers/Cluster/Cluster.scss
@@ -2,6 +2,7 @@
 
 .ydb-cluster {
     --cluster-side-padding: var(--g-spacing-5);
+    --sticky-tabs-height: 40px;
     position: relative;
 
     overflow: auto;
@@ -22,6 +23,7 @@
 
     &__content {
         width: calc(100% - var(--cluster-side-padding));
+        min-height: calc(100% - var(--sticky-tabs-height, 0px));
         //allows controls of TableWithControlsLayout to stick properly
         transform: translateX(var(--cluster-side-padding));
     }
@@ -75,7 +77,6 @@
     }
 
     .ydb-table-with-controls-layout {
-        --sticky-tabs-height: 40px;
         // Total height of all fixed elements above table for sticky header positioning
         --data-table-sticky-header-offset: 102px;
     }
