+++ src/containers/Nodes/PaginatedNodes/index.ts
@@ -1,0 +1,1 @@
+export {PaginatedNodes} from './PaginatedNodes';
